+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/dto/restaurantmenu/RestaurantMenuFormDto.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.dto.restaurant;
+package ProjectDoge.StudentSoup.dto.restaurantmenu;
 
 import ProjectDoge.StudentSoup.entity.restaurant.RestaurantMenuCategory;
 import lombok.Getter;
